extract DataContext interface and create SQL specific implementation
